I took advantage to add support for multiple collection.
The thing is that I have to run --discover after each --configure and 
I am not sure if it is intended to work like this.

TODO:
A. check for ressources and report syncs only if format is supported 
   by the other member(s) in the discover phase
B. Work the event/note/todo sync out
C. Check why todo is reporting XML validation error.



git-svn-id: http://svn.opensync.org/plugins/akonadi-sync/trunk@6144 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e